done uptime, ram, cpu utilization

+++ include/linux_parser.h
@@ -43,7 +43,7 @@ enum CPUStates {
 std::vector<std::string> CpuUtilization();
 long Jiffies();
 long ActiveJiffies();
-long ActiveJiffies(int pid);
+double ActiveJiffies(int pid);
 long IdleJiffies();
 
 // Processes
